Stop generating two types of resolvers
In recent refactor we introduced a new pattern of resolvers which is
better structured and more readable. To keep Gqlgen backward compatible
we started generate two styles of resolvers side by side.

It is now time to sunset the old resolver. This commit removes the old
resolver and update the generation code to use the new resolver
directly.