Handle enum slices as directive parameter
When an enum is defined as an parameter of an directive, the code that is generated contains `[]interface{}` for example `[]interface{}{"ADMIN", "USER"}` instead of the actual type of the enum. This is incompatible with the generated directive function signature which has the explicit type of the enum slice. (see #603)

This change checks after the argument is dumped if the arg is a slice and enum. If this is the case it will replace `[]interface{}` by the actual type.