Add single-file, follow-schema layouts
When `ExecLayout` is set to `follow-schema`, output generated code to a
directory instead of a single file. Each file in the output directory
will correspond to a single *.graphql schema file (plus a
root!.generated.go file containing top-level definitions that are not
specific to a single schema file).

`ExecLayout` defaults to `single-file`, which is the current behavior, so
this new functionality is opt-in.

These layouts expose similar functionality to the `ResolverLayout`s with
the same name, just applied to `exec` instead of `resolver`.

Resolves issue #1265.