Fix the ability of websockets to get errors (#2097)
Because DispatchOperation creates tempResponseContext,
which is passed into Exec, which is then used in _Subscription to
generate the next function. Inside the various subscription functions
when generating next the context was captured there.

Which means later when the returned function from DispatchOperation is
called. The responseContext which accumulates the errors is the
tempResponseContext which we no longer have access to to read the errors
out of it.

Instead add a context to next() so that it can be passed through and
accumulated the errors as expected.

Added a unit test for this as well.

Co-authored-by: Chris Pride <cpride@observeinc.com>