Add @interfaceObject and @composeDirective at Federation 2 directive lists.
Currently, in case of using `@interfaceObject`, in the graphql schema, it generates the directive template at the auto-generated file because those directives are not listed up and not excluded from the generation.

So, it causes "not implemented" error on using @interfaceObject even this directive is federation 2.3 nested directive.

This commit just adding those directives into the list.

Signed-off-by: Hyukjoong Chuck Kim <wangmir@gmail.com>