feat: Implement improved @requires support
The new API for @requires enables you to define resolver functions that compute the value for each field.

It's also type safe — instead of taking in a representations map, we use the known selection set from the fields arg and codegen it into a model just like normal entity resolvers.