feat: Add new data property for operations
To implement a better @requires Federation API, we need to treat the @requires.fields property as a fragment on the current entity and codegen a Go type for it to use as the input to compute the field.

This commit adds a new Operation property that can be populated from plugins. The Federation plugin will populate this with the correct fragments for the selection sets, and we'll generate models for this in modelgen if it exists.