refactor: significantly clean up the federation.gotpl template (#3187)
* fix: fix Federation example

Some configurations weren't working due to a missing resolver.

* chore: Introduce mechanism for running all example Federation subgraphs

This enables engineers to more easily run the debugger on the Federation example. Updated README to show how to use it.

* refactor: significantly clean up the federation.gotpl template

There were a number of inline structs and inline functions that made it extremely hard to reason about what the code is doing. Split these out into smaller functions with less closures and mutation.

Part of the setup to help with fixing #2991.