Upgrading `urfave/cli` to v3
This PR upgrades `urfave/cli` to v3. This has the advantage to get rid of three dependencies, including blackfriday, which is semi-unmaintained and frankly a markdown processor is simply unexpected as a dependency for a gql code-generator.

Unfortunatly, the v3 is still in beta, so you might want to hold actually merging this, once v3 is officially released. At least the changes are prepared then.