Better document how to construct a Server
Expecting a functional constructor; I had a hard time working out how to make a server appropriate for production. This was compounded by all examples using NewDefaultServer. What was not written down is that you should use New and provide properly configured transports, so write that in the deprecation notice.