fix generating enum maps for slices
Previously this would generate the enum lookup maps for the underlying
types of slices, though these functions were not used.

In the case where an enum type is used as an optional field, required
field, and optional array field, the generated maps would be invalid
(missing the appropriate dereference). Because these maps are not
needed, removing them is better.