feat(dataloader): handle thunk resolvers
With how dataloadgen works, we don't need to create as many goroutines, and we
can let it handle all the batching for us, and use their "thunk" function, that
will resolve when the batch of data resolved.

This commit adds a new resolver configuration, "thunk_resolvers", that changes
the return type of resolvers from `(T, error)` to `(func() (T, error), error)`.
This allows us to implement resolvers using `LoadThunk` instead of `Load`, thus
using the whole potential of dataloadgen.

This will end up creating a lot less goroutines in heavy setups that rely on
dataloading, and should greatly improve performance.

This setting could also be enabled for anyone that wants it, without the need
to rely on dataloadgen. Effectively, it gives control to the engineer writing
the implementation of how they handle their goroutines.