fix(modelgen): use ToGoModelName in generateGetters
This is a follow-up from https://github.com/99designs/gqlgen/commit/e371314c3bdf42799f486d091cd8eed06d0e1220

This diff made it so that we use `templates.ToGoModelName` instead of
`templates.ToGo`. This should also be done in the modelgen functions, as if we
enrich a model with interfaces that generate more getters, those will never use
the model name with the number appended to it.