feat(resolveField): allow field middlewares to return a marshaler directly (#3928)
In some specific cases, we'd like the ability from a field resolver to return a
marshaler directly. This is needed for specific performance-sensitive codepaths
we have in our application, where part of the request might already be resolved
and we don't want to resolve it again, as the performance hit from doing so
would be too high.

This shouldn't impact any already existing usecase, and only make those field
middlewares more capable.